% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate.R
\name{NilsEstimate}
\alias{NilsEstimate}
\alias{NilsEstimateBalanced}
\title{Estimate the total}
\usage{
NilsEstimate(
  plot_data,
  tract_data,
  psus,
  category_psu_map,
  area = 46390172,
  tract_area = 196 * 100 * pi
)

NilsEstimateBalanced(
  plot_data,
  tract_data,
  auxiliaries,
  psus,
  category_psu_map,
  area = 46390172,
  tract_area = 196 * 100 * pi,
  size_of_neighbourhood = NULL
)
}
\arguments{
\item{plot_data}{A data frame containing information about observations.}

\item{tract_data}{A matrix containing tract ids and the psus they belong to.}

\item{psus}{An ordered vector of psus, ranging from largest to smallest.}

\item{category_psu_map}{A matrix containing categories and the psus they map to.}

\item{area}{The size of the area frame.}

\item{tract_area}{The area of a tract.}

\item{auxiliaries}{A (double) matrix of auxiliary information used in the balancing.
Must match the size (and order) of tract_data.}
}
\value{
A NilsEstimate object (list).
}
\description{
Estimates the total of some variable, surveyed using the NILS hiearchical design framework.
}
\details{
\code{category_psu_map} contains information about the categories used in the
design.
The object should be a matrix with the following columns (in order):
\itemize{
\item the category id (integer), as used in \code{plot_data};
\item the psu id (integer) of the largest psu that the category begins to be
sampled in.
}

\code{tract_data} contains information about all sampled tracts, even those where
no interesting category was found.
The object should be a matrix with the following columns (in order):
\itemize{
\item the tract id (integer) of the sampled tracts;
\item the psu id (integer) of the psu that the tract were sampled in.
}

\code{plot_data} contains information about observed values on plot level. The
object can contain multiple records per plot.
The object should be a data frame with the following columns (in order):
\itemize{
\item the tract id (integer) of the parent tract;
\item the category id (integer) of the category recorded for the plot;
\item the design weight (double) for the plot, conditioned on the tract;
\item the value of the target variable (double).
}

The functions returns a NilsEstimate object (list) with the following names:
\itemize{
\item \code{estimate}: the estimate of the total;
\item \code{variance}: the variance estimate of the total estimate;
\item \code{cat_estimates}: the estimates per category;
\item \code{cat_covmat}: the estimated covariance matrix of the per category estimates.
}

\subsection{NilsEstimateBalanced}{

For the balanced variant, the variance is calculated by using a local
neighbourhood deviance measure.
The size of this local neighbourhood defaults to 4 for the smallest PSUs, and
increases linearly by size.

As the covariance between two categories belonging to different psus are
measured on the intersect between these categories, e.g. on the smaller of
the psus, the smaller psu also decides the size of the local neighbourhood.

It is possible to provide custom neighbourhood sizes by adding a third column
to \code{psu_sizes}:
\itemize{
\item size of neighbourhood
}
}
}
