% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nilsier-package.R
\docType{package}
\name{nilsier}
\alias{nilsier}
\alias{nilsier-package}
\title{nilsier: Design-based estimators for NILS}
\description{
Provides estimators for NILS hierarchical design.
}
\author{
Wilmer Prentius \email{wilmer.prentius@slu.se}.
}
\keyword{internal}
