% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-utils.R
\name{vcov.NilsEstimate}
\alias{vcov.NilsEstimate}
\title{Covariance matrix for NILS category estimates}
\usage{
\method{vcov}{NilsEstimate}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{A \code{NilsEstimate} object.}

\item{complete}{Logical. If \code{FALSE}, excludes apparent zero-tracts.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Returns the covariance matrix of a \code{NilsEstimate} object.
}
\examples{
obj = NilsEstimate(plots, tracts, psus, category_psu_map);
vcov(obj);

}
