% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-utils.R
\name{coef.NilsEstimate}
\alias{coef.NilsEstimate}
\alias{vcov.NilsEstimate}
\title{Estimates of the NilsEstimate}
\usage{
\method{coef}{NilsEstimate}(obj)

\method{vcov}{NilsEstimate}(obj, complete = TRUE)
}
\arguments{
\item{obj}{An object of class \code{NilsEstimate}}

\item{complete}{If \code{FALSE}, excludes apparent 0-tracts}
}
\description{
Returns a vector of the estimate and variance estimate of the NilsEstimate object.

Returns the covariance matrix of the NilsEstimate object
}
