% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate.R
\name{NilsEstimate}
\alias{NilsEstimate}
\alias{NilsEstimateBalanced}
\title{Estimate the total}
\usage{
NilsEstimate(
  plot_data,
  tract_data,
  psus,
  category_psu_map,
  area = 46519242.1175867,
  tract_area = 196 * 100 * pi
)

NilsEstimateBalanced(
  plot_data,
  tract_data,
  auxiliaries,
  psus,
  category_psu_map,
  area = 46519242.1175867,
  tract_area = 196 * 100 * pi,
  size_of_neighbourhood = NULL
)
}
\arguments{
\item{plot_data}{A data frame containing information about observations.}

\item{tract_data}{A matrix containing tract ids and the psus they belong to.}

\item{psus}{An ordered vector of PSUs, ranging from largest to smallest (in terms of PSU sizes).}

\item{category_psu_map}{A matrix containing categories and the psus they map to.}

\item{area}{The size of the area frame. The area frame is most likely larger than the area of
interest.}

\item{tract_area}{The area of a tract, in the same unit as the value of the target variable.}

\item{auxiliaries}{A (double) matrix of auxiliary information used in the balancing.
Must match the size (and order) of tract_data.}

\item{size_of_neighbourhood}{An optional vector of sizes of the neighbourhood for each PSU.}
}
\value{
A \code{NilsEstimate} object, essentially a data frame with the estimates per category. The
data frame has the following columns
\describe{
\item{Cat. ID}{Category id number}
\item{Est. total}{The estimated total within the category}
\item{Est. variance}{The estimated variance of the estimator of total within the category}
\item{Positive tracts}{The number of tracts on which there was a recorded positive value of the
variable of interest, within the category}
}
}
\description{
Estimates the total of some variable, surveyed using the NILS hiearchical design framework.
}
\details{
\code{category_psu_map} contains information about the categories used in the design.
The object should be a matrix with the following columns (in order):
\enumerate{
\item The category id number (integer), as used in \code{plot_data}.
\item The PSU id number (integer) of that the category belongs to; the smallest PSU (id) in which the
category is sampled.
}

\code{tract_data} contains information about all sampled tracts, even those where no interesting
category was found.
The object should be a matrix with the following columns (in order):
\enumerate{
\item The tract id numbers (integer) of all sampled tracts.
\item The PSU id number (integer) of the PSU that last sampled the tract; the smallest PSU which
contains the tract.
}

\code{plot_data} contains information about observed values on plot level.
The object can contain multiple records per plot.
The object should be a data frame with the following columns (in order):
\enumerate{
\item the tract id number (integer) of the parent tract.
\item the category id number (integer) of the category recorded for the plot.
\item the design weight (double) for the plot, conditioned on the tract.
\item the value of the target variable (double).
}

\subsection{NilsEstimateBalanced}{

For the balanced variant, the variance is calculated by using a local neighbourhood deviance
measure.
The size of this local neighbourhood defaults to 4 for the smallest PSUs, and increases linearly
by size.

As the covariance between two categories belonging to different PSUs are measured on the
intersect between these categories, e.g. on the smaller of the PSUs, the smaller PSU also decides
the size of the local neighbourhood.

It is possible to provide a matrix as \code{psus} instead of the vector. The matrix should have the
following columns:
\enumerate{
\item The PSU IDs, ranging from largest to smallest (in terms of PSU sizes).
\item The size of the neighbourhoods to use for each PSU.
}

If a matrix is not provided as \code{psus}, the neighbourhood size of PSU $k$ will default to
\deqn{4 \frac{n_{k}}{n_{(0)}} ,}
where \eqn{n_{k}} is the size of PSU \eqn{k}, and \eqn{n_{(0)}} is the size of the smallest PSU.
}
}
\examples{
\dontrun{
obj = NilsEstimate(plots, tracts, psus, category_psu_map);
}

\dontrun{
obj = NilsEstimateBalanced(plots, tracts, tract_auxilliaries, psus, category_psu_map);
}

}
