% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate.R
\name{PreparePlotData}
\alias{PreparePlotData}
\title{Prepare plot data}
\usage{
PreparePlotData(data, tid, cat, dw, y)
}
\description{
Takes the column names as formulas, and returns a correctly ordered data frame.
}
