% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-print.R
\name{print.NilsEstimate}
\alias{print.NilsEstimate}
\title{Print}
\usage{
\method{print}{NilsEstimate}(x, complete = TRUE, ...)
}
\arguments{
\item{x}{the NilsEstimate object}

\item{complete}{If \code{FALSE}, excludes apparent 0-tracts}

\item{...}{additional unused arguments}
}
\description{
Print
}
