% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-print.R
\name{print.NilsEstimate}
\alias{print.NilsEstimate}
\title{Print}
\usage{
\method{print}{NilsEstimate}(obj, complete = TRUE)
}
\arguments{
\item{complete}{If \code{FALSE}, excludes apparent 0-tracts}
}
\description{
Print
}
