% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-print.R
\name{print.NilsEstimate}
\alias{print.NilsEstimate}
\title{Print a NILS estimate}
\usage{
\method{print}{NilsEstimate}(x, complete = TRUE, ...)
}
\arguments{
\item{x}{A \code{NilsEstimate} object.}

\item{complete}{Logical. If \code{FALSE} (default), excludes apparent zero-tracts from the printed
output.}

\item{...}{Additional arguments (currently unusued)}
}
\value{
Invisibly returns the input \code{NilsEstimate} object.
}
\description{
Prints a summary of a \code{NilsEstimate} object.
}
