% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-summary.R
\name{summary.NilsEstimate}
\alias{summary.NilsEstimate}
\alias{print.summary.NilsEstimate}
\title{Summary of the NilsEstimate}
\usage{
\method{summary}{NilsEstimate}(obj)

\method{print}{summary.NilsEstimate}(obj)
}
\arguments{
\item{obj}{An object of class \code{NilsEstimate}}
}
\description{
Returns a list of summary statistics of the NilsEstimate object
}
\details{
The returned list has the following names:
\itemize{
\item \code{estimate} for the estimate of the total,
\item \code{variance} for the variance estimate of the total,
\item \code{rel_se} for the relative standard error of the total.
}

If the data has not been filtered, it also contains:
\itemize{
\item \code{nonnil_tracts} for the number of non-nil tracts
}
}
