% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NilsEstimate-summary.R
\name{summary.NilsEstimate}
\alias{summary.NilsEstimate}
\alias{print.summary.NilsEstimate}
\title{Summary of the NilsEstimate}
\usage{
\method{summary}{NilsEstimate}(object, ...)

\method{print}{summary.NilsEstimate}(x, ...)
}
\arguments{
\item{object, x}{An object of class \code{NilsEstimate}}

\item{...}{additional unused arguments}
}
\description{
Returns a list of summary statistics of the NilsEstimate object
}
\details{
The returned list has the following names:
\describe{
\item{estimate}{the estimate of the total}
\item{variance}{the estimated variance of the estimator of the total}
\item{rel_se}{the estimated relative standard error of the estimator of the total}
\item{nonnil_tracts}{the number of non-nil tracts (not shown if filter has been applied)}
}
}
\examples{
\dontrun{
obj = NilsEstimate(plots, tracts, psus, category_psu_map);
summary(obj);
}

}
